Add initial zsh support

+++ setup.sh
@@ -26,10 +26,22 @@ brew install autojump
 # md5sum
 brew install md5sha1sum
 
+# bash vs zsh
+read -p "Configure (b) bash (default) or (z) oh-my-zsh? " -n 1 -r
+echo
+if [[ $REPLY =~ ^[zZ]$ ]]
+then
+    brew install zsh zsh-completions
+    chsh -s /bin/zsh
+    sh -c "$(curl -fsSL https://raw.github.com/robbyrussell/oh-my-zsh/master/tools/install.sh)"
+    ln -f .zshrc ~/.zshrc
+else
+    ln -f .bash_profile ~/.bash_profile
+fi
+
 # Creating hard links to the home folder; using -f to replace existing links
 ln -f .git-prompt.sh ~/.git-prompt.sh
 ln -f .git-completion.bash ~/.git-completion.bash
-ln -f .bash_profile ~/.bash_profile
 ln -f .gitconfig ~/.gitconfig
 ln -f gradle.properties ~/.gradle
 ln -f .gradle-tab-completion.bash ~/.gradle-tab-completion.bash
